requirement.txt
langchain
langchain_community
openai
flask
python-dotenv